Extract editor logic into its own component
